code/requirements.txt
Django==3.0.2
djangorestframework~=3.11.0
factory-boy==3.2.0
psycopg2-binary~=2.8.4
python-decouple~=3.3
flake8==3.8.2
isort==4.3.21
black==20.8b1
mypy==0.770